Improved instructions for Ubuntu installation
Adding instructions for installing on Ubuntu by PPA. This should be stable, I tested it on 3 VMs and two real machines. The package currently isn't set up to allow the use of avconv (which I may, or may not fix later) and 12.04 doesn't have a new enough version of ffmpeg to support some of x11grab's options, so ppa:jon-severinsson/ffmpeg is needed. ppa:webupd8team/y-ppa-manger provides yad. The last PPA provides silentcast (that should be obvious ;). I will actively maintain it for as long as I can.